Move LambdaLifting to separate file
